cleanup packing method
