+++ scratch.js
@@ -1,23 +1,12 @@
 'use strict';
-// const arr = require('./test/maleNames');
+let arr = require('./test/maleNames');
 // const trieHard = require('./bits');
 const trieHard = require('./src');
 
-var arr = [
-  'aaa',
-  'aa',
-  'aaa',
-  'aaa',
-  'aaab',
-  'a',
-  'aaaa',
-  'b',
-  'b22',
-// '35',
-// '1234e',
-// '567890',
-// 'fe567890',
-// 'workin\'',
+arr = [
+  'fun',
+  'fung',
+  'fun2',
 ];
 
 // var words = [
